+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+### Changed
+- [Release] Move React button `full` size option to its own prop for `width`. (#615)
 
 ## [18.3.0][18.3.0] - 2016-10-07
 ### Added